requirements_cmms.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
APScheduler==3.10.4
cryptography==41.0.7
requests==2.31.0
Werkzeug==3.0.0
openpyxl==3.1.2
